Refactor common elements into a module
