+++ app/services/ect_at_school_periods/switch_training.rb
@@ -4,6 +4,7 @@ module ECTAtSchoolPeriods
 
   class SwitchTraining
     include TrainingPeriodSources
+    include Schedules::Reuse
 
     def self.to_school_led(...) = new(...).to_school_led
     def self.to_provider_led(...) = new(...).to_provider_led
@@ -87,37 +88,15 @@ module ECTAtSchoolPeriods
       end
     end
 
-    def schedule
-      existing_schedule if reuse_existing_schedule?
-    end
-
     def reuse_existing_schedule?
       return false unless @ect_at_school_period.school_led_training_programme?
-      return false unless existing_schedule
-      return false if contract_period_reassignment_required?
 
-      existing_schedule.contract_period != contract_period_at_transition
+      super
     end
-
-    def existing_schedule
-      last_provider_led_training_period&.schedule
-    end
-
-    def date_of_transition = [@ect_at_school_period.started_on, Date.current].max
 
     def contract_period_reassignment_required?
-      @ect_at_school_period.school_led_training_programme? && contract_period_reassignment.required?
+      @ect_at_school_period.school_led_training_programme? && super
     end
-
-    def contract_period_reassignment
-      @contract_period_reassignment ||= ContractPeriods::Reassignment.new(training_period: last_provider_led_training_period)
-    end
-
-    def last_provider_led_training_period
-      @last_provider_led_training_period ||= @ect_at_school_period.training_periods.provider_led_training_programme.latest_first.first
-    end
-
-    delegate :successor_contract_period, to: :contract_period_reassignment
 
     def set_ect_payments_frozen_year!
       return unless contract_period_reassignment_required?
@@ -125,10 +104,6 @@ module ECTAtSchoolPeriods
       @ect_at_school_period
         .teacher
         .update!(ect_payments_frozen_year: contract_period_reassignment.assigned_contract_period.year)
-    end
-
-    def contract_period_at_transition
-      @contract_period_at_transition ||= ContractPeriod.containing_date(date_of_transition)
     end
 
     def finish_training_period!
@@ -204,6 +179,6 @@ module ECTAtSchoolPeriods
       @mentor_at_school_period.teacher.mentor_became_ineligible_for_funding_on.present?
     end
 
-    def started_on = date_of_transition
+    def period = @ect_at_school_period
   end
 end
