+++ app/services/mentor_at_school_periods/change_lead_provider.rb
@@ -18,5 +18,7 @@ module MentorAtSchoolPeriods
         author:
       ).finish!
     end
+
+    def contract_period_reassignment_required? = false
   end
 end
